pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-sqlalchemy>=3.1.1",
    "groq>=1.1.2",
    "gunicorn>=25.3.0",
    "pdfplumber>=0.11.9",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "werkzeug>=3.1.8",
]
